Füge defineComponent-Funktion hinzu und modularisiere Komponentenlogik in Submodule
