+++ CHANGELOG.md
@@ -35,11 +35,13 @@ and this project adheres to Semantic Versioning.
   - New helpers: `animate`, `sequence`, `timeline`, `scrollAnimate`, `stagger`, `flipElements`.
   - New presets: `easingPresets`, `keyframePresets`, plus individual easing exports.
   - Improved reduced-motion support via `prefersReducedMotion()`.
+- **Component**: `defineComponent()` factory for manual class creation and custom registration.
 
 ### Changed (Unreleased)
 
 - **Security**: Internals modularized (sanitize core, Trusted Types, CSP helpers, constants/types) with no API changes.
 - **Router**: Internals modularized into focused submodules with no public API changes.
+- **Component**: Internals modularized into focused submodules with no public API changes.
 
 ### Fixed (Unreleased)
